Refactor SCSS and file structure.
