--- src/components/FilterComponent/RoleFilter/RoleFilter.scss
@@ -1,73 +1,0 @@
-.role-filter {
-    .filter-section {
-      border-bottom: 1px solid #e0e0e0;
-      padding-bottom: 15px;
-      margin-bottom: 15px;
-      cursor: pointer;
-      transition: all 0.3s ease;
-  
-      &.collapsed {
-        max-height: 40px;
-        overflow: hidden;
-      }
-  
-      h3 {
-        font-size: 1em;
-        color: rgb(255, 255, 255);
-        margin-bottom: 10px;
-      }
-  
-      div {
-        margin-bottom: 10px;
-  
-        input[type='checkbox'] {
-          display: none;
-  
-          + label {
-            position: relative;
-            padding-left: 24px;
-            cursor: pointer;
-  
-            &:before {
-              content: '';
-              position: absolute;
-              left: 0;
-              top: 50%;
-              transform: translateY(-50%);
-              width: 16px;
-              height: 16px;
-              border: 2px solid #000000;
-              border-radius: 3px;
-              background-color: #fff;
-              transition: all 0.3s ease;
-            }
-  
-            &:after {
-              content: '';
-              position: absolute;
-              left: 4px;
-              top: 50%;
-              transform: translateY(-50%) scale(0);
-              width: 8px;
-              height: 16px;
-              border-right: 2px solid #0b75ff;
-              border-top: 2px solid #0b75ff;
-              transition: transform 0.2s ease;
-              transform-origin: bottom left;
-              rotate: 45deg;
-            }
-          }
-  
-          &:checked + label {
-            &:before {
-              border-color: #009688;
-            }
-            &:after {
-              transform: translateY(-50%) scale(1);
-            }
-          }
-        }
-      }
-    }
-  }
-
